refactoring

+++ gl-research-toolkit/scene/include/glrt/scene/resources/utilities/gl-texture.h
@@ -120,7 +120,7 @@ public:
 struct GlTexture::TextureAsFloats
 {
   UncompressedImage image;
-  QVector<byte> textureData;
+  QVector<float> textureData;
   byte* data = nullptr;
   quint32 width = 0, components_per_row=0, height = 0, depth = 0, rowCount=0;
